Fix HTML5 mode (needed to let facebook properly link to our pages)
attn @mrpollo - the built in webserver in the grunt task hasn't yet
been changed to return index.html for all invalid URLs (needed so angular
does the right thing on page reloads) - I have changed S3 to do the
right thing.

This means that the auto page reload in the local server is busted, you
need to follow links inside the app instead.  I'm working on a fix for
this problem.